Create AImate
The academic validation pressure on students is no more an ignorable thing since many of the students ended up their precious lives because of not knowing how to handle all the things. As an University student most of my friends and I are can not be able to concentrate simultaneously on curriculum and extra curriculum eventually this makes us get exhausted sometimes by overthinking and even make as anxious, stressed and depressed. We also should focus on our mental and physical health at the same time being connected with friends could help us to feel more accompanied. The students need a good connection with family and friends, they need to exercise on daily basis to elongate their lifetime and to elevate their self-discipline, the also should be guided with a perfect planner if they do not have time planning the study. So, the AI tool I have generated does all these things integrated. I consider my project will be helpful for the students who are all facing challenges in these such things also I planned to add features and develop this as a super useful product for Student community hopefully in the future.